Fix narrowing conversions when iterating inside the search box

+++ CLUEstering/include/Clustering.h
@@ -150,10 +150,10 @@ private:
       // loop over bins in the search box
       for (int binId : search_box) {
         // get the size of this bin
-        int binSize = tiles[binId].size();
+        size_t binSize = tiles[binId].size();
 
         // iterate inside this bin
-        for (int binIter{}; binIter < binSize; ++binIter) {
+        for (size_t binIter{}; binIter < binSize; ++binIter) {
           int j{tiles[binId][binIter]};
           // query N_{dc_}(i)
           float dist_ij{distance(i, j)};
@@ -198,10 +198,10 @@ private:
       // loop over all bins in the search box
       for (int binId : search_box) {
         // get the size of this bin
-        int binSize{tiles[binId].size()};
+        size_t binSize{tiles[binId].size()};
 
         // iterate inside this bin
-        for (int binIter{}; binIter < binSize; ++binIter) {
+        for (size_t binIter{}; binIter < binSize; ++binIter) {
           int j{tiles[binId][binIter]};
           // query N'_{dm}(i)
           bool foundHigher{(points_.rho[j] > rho_i)};
